+++ Scribe/Scribe-Core/src/utopia/scribe/core/controller/logging/ScribeLike.scala
@@ -4,7 +4,7 @@ import utopia.flow.generic.model.immutable.{Model, Value}
 import utopia.flow.operator.ScopeUsable
 import utopia.flow.util.logging.Logger
 import utopia.scribe.core.model.enumeration.Severity
-import utopia.scribe.core.model.enumeration.Severity.{Critical, Debug, Info, Recoverable, Unrecoverable, Warning}
+import utopia.scribe.core.model.enumeration.Severity._
 
 /**
   * Common trait for loggers on both the client and the server side
@@ -34,14 +34,6 @@ trait ScribeLike[+Repr] extends Logger with ScopeUsable[Repr]
 	  * @return Copy of this logger that logs the specified issue variant
 	  */
 	def variant(details: Model): Repr
-	
-	/**
-	  * Logs an error
-	  * @param error The error to log (optional)
-	  * @param message Additional message to log (may be empty)
-	  * @param details Additional details to log (may be empty)
-	  */
-	def apply(error: Option[Throwable], message: String, details: Model): Unit
 	
 	
 	// COMPUTED -----------------------------
@@ -80,11 +72,6 @@ trait ScribeLike[+Repr] extends Logger with ScopeUsable[Repr]
 	def critical = apply(Critical)
 	
 	
-	// IMPLEMENTED  -------------------------
-	
-	override def apply(error: Option[Throwable], message: String): Unit = apply(error, message, Model.empty)
-	
-	
 	// OTHER    -----------------------------
 	
 	/**
@@ -99,24 +86,4 @@ trait ScribeLike[+Repr] extends Logger with ScopeUsable[Repr]
 	  * @return Copy of this logger that includes the specified detail in logging entries
 	  */
 	def variant(key: String, detail: Value): Repr = variant(Model.from(key -> detail))
-	
-	/**
-	  * Logs an error
-	  * @param error The error to log
-	  * @param message Additional error message to record (optional)
-	  * @param details Details about this specific issue occurrence (optional)
-	  */
-	def apply(error: Throwable, message: String = "", details: Model = Model.empty): Unit =
-		apply(Some(error), message, details)
-	/**
-	  * Logs a message with details
-	  * @param message Message to log
-	  * @param details Details to include
-	  */
-	def apply(message: String, details: Model): Unit = apply(None, message, details)
-	/**
-	  * Logs an entry with no message but some details instead
-	  * @param details Details to log
-	  */
-	def apply(details: Model): Unit = apply(None, "", details)
 }