+++ Courier/Changes.md
@@ -1,5 +1,8 @@
 # Utopia Courier - List of Changes
 
+## v1.1.7 (in development)
+Supports **Flow v2.9**
+
 ## v1.1.6 - 15.03.2026
 Supports **Flow v2.8**